+++ tests/Unit/Authentication/NativeAuthenticatorTest.php
@@ -19,11 +19,13 @@ declare(strict_types=1);
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-namespace Youthweb\Api\Tests\Unit;
+namespace Youthweb\Api\Tests\Unit\Authentication;
 
-use Youthweb\Api\YouthwebAuthenticator;
+use PHPUnit\Framework\TestCase;
+use Youthweb\Api\Authentication\NativeAuthenticator;
+use Youthweb\OAuth2\Client\Provider\Youthweb;
 
-class YouthwebAuthenticatorTest extends \PHPUnit\Framework\TestCase
+class YouthwebAuthenticatorTest extends TestCase
 {
     /**
      * Create a Authenticator with mocks of all collaborators
@@ -35,12 +37,12 @@ class YouthwebAuthenticatorTest extends \PHPUnit\Framework\TestCase
         $options = array_merge($default_options, $options);
 
         $default_collaborators = [
-            'oauth2_provider' => $this->createMock('Youthweb\OAuth2\Client\Provider\Youthweb'),
+            'oauth2Provider' => $this->createMock(Youthweb::class),
         ];
 
         $collaborators = array_merge($default_collaborators, $collaborators);
 
-        return new YouthwebAuthenticator($options, $collaborators);
+        return new NativeAuthenticator($options, $collaborators);
     }
 
     /**
@@ -48,16 +50,16 @@ class YouthwebAuthenticatorTest extends \PHPUnit\Framework\TestCase
      */
     public function testGetAuthorizationUrlReturnsUrl(): void
     {
-        $oauth2_provider = $this->createMock('Youthweb\OAuth2\Client\Provider\Youthweb');
+        $oauth2Provider = $this->createMock(Youthweb::class);
 
         $url = 'https://example.org';
 
-        $oauth2_provider->expects($this->once())
+        $oauth2Provider->expects($this->once())
             ->method('getAuthorizationUrl')
             ->willReturn($url);
 
         $authenticator = $this->createAuthenticator([], [
-            'oauth2_provider' => $oauth2_provider,
+            'oauth2Provider' => $oauth2Provider,
         ]);
 
         $this->assertSame($url, $authenticator->getAuthorizationUrl());
@@ -68,16 +70,16 @@ class YouthwebAuthenticatorTest extends \PHPUnit\Framework\TestCase
      */
     public function testGetStateReturnsState(): void
     {
-        $oauth2_provider = $this->createMock('Youthweb\OAuth2\Client\Provider\Youthweb');
+        $oauth2Provider = $this->createMock(Youthweb::class);
 
         $state = 'random_string';
 
-        $oauth2_provider->expects($this->once())
+        $oauth2Provider->expects($this->once())
             ->method('getState')
             ->willReturn($state);
 
         $authenticator = $this->createAuthenticator([], [
-            'oauth2_provider' => $oauth2_provider,
+            'oauth2Provider' => $oauth2Provider,
         ]);
 
         $this->assertSame($state, $authenticator->getState());
@@ -88,16 +90,20 @@ class YouthwebAuthenticatorTest extends \PHPUnit\Framework\TestCase
      */
     public function testGetStateWorkaroundReturnsState(): void
     {
-        $oauth2_provider = $this->createMock('Youthweb\OAuth2\Client\Provider\Youthweb');
+        $oauth2Provider = $this->createMock(Youthweb::class);
+
+        $oauth2Provider->expects($this->once())
+            ->method('getAuthorizationUrl')
+            ->willReturn('');
 
         $state = 'random_string';
 
-        $oauth2_provider->expects($this->exactly(2))
+        $oauth2Provider->expects($this->exactly(2))
             ->method('getState')
             ->will($this->onConsecutiveCalls('', $state));
 
         $authenticator = $this->createAuthenticator([], [
-            'oauth2_provider' => $oauth2_provider,
+            'oauth2Provider' => $oauth2Provider,
         ]);
 
         $this->assertSame($state, $authenticator->getState());
@@ -121,21 +127,21 @@ class YouthwebAuthenticatorTest extends \PHPUnit\Framework\TestCase
      */
     public function testGetAccessTokenWithAuthCodeAndStateSavesToken(): void
     {
-        $oauth2_provider = $this->createMock('Youthweb\OAuth2\Client\Provider\Youthweb');
+        $oauth2Provider = $this->createMock(Youthweb::class);
         $access_token = $this->createMock('League\OAuth2\Client\Token\AccessToken');
 
-        $oauth2_provider->expects($this->once())
+        $oauth2Provider->expects($this->once())
             ->method('getAccessToken')
             ->willReturn($access_token);
 
         $authenticator = $this->createAuthenticator(
             [
-                'client_id'     => 'client_id',
-                'client_secret' => 'client_secret',
-                'redirect_url'  => 'https://example.org/callback',
+                'clientId'     => 'client_id',
+                'clientSecret' => 'client_secret',
+                'redirectUrl'  => 'https://example.org/callback',
             ],
             [
-                'oauth2_provider' => $oauth2_provider,
+                'oauth2Provider' => $oauth2Provider,
             ]
         );